Unify the two EnumerableExtensions class files.
